add extra check for id-contacts

+++ controllers/contacts.js
@@ -13,6 +13,10 @@ const listContacts = async (req, res) => {
 };
 
 const getContactById = async (req, res) => {
+  const { _id: owner } = req.user;
+  if (req.user._id !== { owner }) {
+    throw HttpError(404, "Not found");
+  }
   const { contactId } = req.params;
   const result = await Contact.findById(contactId);
   res.json({ result });
@@ -25,6 +29,10 @@ const addContact = async (req, res) => {
 };
 
 const deleteContact = async (req, res) => {
+  const { _id: owner } = req.user;
+  if (req.user._id !== { owner }) {
+    throw HttpError(404, "Not found");
+  }
   const { contactId } = req.params;
   const result = await Contact.findByIdAndDelete(contactId);
   if (!result) {
@@ -36,6 +44,10 @@ const deleteContact = async (req, res) => {
 };
 
 const updateContact = async (req, res) => {
+  const { _id: owner } = req.user;
+  if (req.user._id !== { owner }) {
+    throw HttpError(404, "Not found");
+  }
   const { contactId } = req.params;
   const result = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
@@ -47,6 +59,10 @@ const updateContact = async (req, res) => {
 };
 
 const updateFavorite = async (req, res) => {
+  const { _id: owner } = req.user;
+  if (req.user._id !== { owner }) {
+    throw HttpError(404, "Not found");
+  }
   const { contactId } = req.params;
   const result = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
